ccube/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-crispy-forms==1.9.2
django-currentuser==0.5.2
django-nested-admin==3.3.2
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
